# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "AFSE API",
    "description": "API for Managing AFSE",
    "contact": {
      "name": "API Support",
      "email": "matteo.martinelli5@studenti.unimi.it"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/info": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Recupera le info utente",
        "description": "Ritorna le informazioni relative al profilo dell utente autenticato. Richiede un JWT token passato tramite cookie.",
        "responses": {
          "200": {
            "description": "Profilo utente recuperato con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "oid": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "L ID dell utente"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "66dc65c5da63d679f7133e3c"
                                }
                              }
                            }
                          },
                          "required": [
                            "oid"
                          ]
                        }
                      }
                    },
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Nome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "surname": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Cognome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "username": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Username dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g@example.com"
                        }
                      }
                    },
                    "pwd": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Password hash dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "cd0aa9856147b6c5b4ff2b7dfee5da20aa38253099ef1b4a64aced233c9afe29"
                        }
                      }
                    },
                    "isHero": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "boolean"
                        },
                        "description": {
                          "type": "string",
                          "example": "Indica se l utente è un supereroe"
                        },
                        "example": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "favouriteHero": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "favouriteHeroId": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "integer"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "ID del supereroe preferito"
                                },
                                "example": {
                                  "type": "number",
                                  "example": 1011334
                                }
                              }
                            },
                            "favouriteHeroName": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Nome del supereroe preferito"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "Spider-Man"
                                }
                              }
                            },
                            "favouriteHeroImage": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "URL dell immagine del supereroe preferito"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "http://i.annihil.us/u/prod/marvel/i/mg/9/c0/527bb7b37ff55.jpg"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "profilePicture": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "URL dell immagine del profilo dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "http://i.annihil.us/u/prod/marvel/i/mg/9/c0/527bb7b37ff55.jpg"
                        }
                      }
                    },
                    "coinAmount": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "description": {
                          "type": "string",
                          "example": "Numero di monete dell utente"
                        },
                        "example": {
                          "type": "number",
                          "example": 9
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Utente non autenticato o JWT token mancante"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Recupera il profilo utente",
        "description": "Ritorna le informazioni relative al profilo dell utente autenticato. Richiede un JWT token passato tramite cookie.",
        "responses": {
          "200": {
            "description": "Profilo utente recuperato con successo",
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "oid": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "L ID dell utente"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "66dc65c5da63d679f7133e3c"
                                }
                              }
                            }
                          },
                          "required": [
                            "oid"
                          ]
                        }
                      }
                    },
                    "name": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Nome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "surname": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Cognome dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "username": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Username dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g"
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Email dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "g@example.com"
                        }
                      }
                    },
                    "pwd": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "Password hash dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "cd0aa9856147b6c5b4ff2b7dfee5da20aa38253099ef1b4a64aced233c9afe29"
                        }
                      }
                    },
                    "isHero": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "boolean"
                        },
                        "description": {
                          "type": "string",
                          "example": "Indica se l utente è un supereroe"
                        },
                        "example": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "favouriteHero": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        },
                        "properties": {
                          "type": "object",
                          "properties": {
                            "favouriteHeroId": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "integer"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "ID del supereroe preferito"
                                },
                                "example": {
                                  "type": "number",
                                  "example": 1011334
                                }
                              }
                            },
                            "favouriteHeroName": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "Nome del supereroe preferito"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "Spider-Man"
                                }
                              }
                            },
                            "favouriteHeroImage": {
                              "type": "object",
                              "properties": {
                                "type": {
                                  "type": "string",
                                  "example": "string"
                                },
                                "description": {
                                  "type": "string",
                                  "example": "URL dell immagine del supereroe preferito"
                                },
                                "example": {
                                  "type": "string",
                                  "example": "http://i.annihil.us/u/prod/marvel/i/mg/9/c0/527bb7b37ff55.jpg"
                                }
                              }
                            }
                          }
                        }
                      }
                    },
                    "profilePicture": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "URL dell immagine del profilo dell utente"
                        },
                        "example": {
                          "type": "string",
                          "example": "http://i.annihil.us/u/prod/marvel/i/mg/9/c0/527bb7b37ff55.jpg"
                        }
                      }
                    },
                    "coinAmount": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "description": {
                          "type": "string",
                          "example": "Numero di monete dell utente"
                        },
                        "example": {
                          "type": "number",
                          "example": 9
                        }
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Utente non autenticato o JWT token mancante"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "jwt",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT token passato tramite cookie"
      }
    },
...
  }
}